requirements.txt
asgiref==3.7.2
astroid==3.0.1
black==23.11.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
Django==4.2.7
djangorestframework==3.14.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
PyJWT==2.8.0
pylint==3.0.2
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.2
tzdata==2023.3
